api/requirements.txt
line-bot-sdk==1.21.0
Flask==2.1.1
requests==2.27.1
googletrans==4.0.0-rc1
PyPDF2==1.26.0
fpdf==1.7.2
